+++ app/src/main/java/woowacourse/payments/ui/cardlist/components/CardListTopBar.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.cardlist
+package woowacourse.payments.ui.cardlist.components
 
 import androidx.compose.material3.CenterAlignedTopAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
